close #14
Introduced a new non-unique displayName field in users model.
For local auth strategy it equals the username.
For oAuth2 providers, it is the username (or displayName) in the external system, while the actual username is prepended with a prefix to identify the oAuth2 provider and thus making the field unique